US125152 Ignore additional values
Came up while attempting to parse a Siren response that has extra
values, in particular a min/max on a Field. Siren doesn't really make a
decision about extra values one way or the other, so at the very least
we shouldn't be exploding when they are present. (Ideally, the values
would get passed through, but that's difficult in strongly-typed
languages.)

This adds a case to simply skip a token if it doesn't match our expected
set of token names. This does mean an Entity/Link/Action/Field parsed
with KSiren will _not_ fully represent the original JSON string, but
this is a slight improvement over just breaking.

Note that this does mean `fromJson` is now lossy, but only in cases
where we previously would have just thrown an exception anyway.